Move utility functions from api/ into utils/
