+++ src/brain/gocdNoDeploysAlert/index.test.ts
@@ -17,7 +17,7 @@ import { db } from '@utils/db';
 
 import { DEPLOYS_FAILING_LIMIT_MS, gocdNoDeploysAlert, handler } from '.';
 
-jest.mock('@api/getUser');
+jest.mock('@/utils/github/getUser');
 
 describe('gocdSlackFeeds', function () {
   let fastify: Fastify;
