+++ src/brain/gocdSlackFeeds/deployFeed.ts
@@ -10,7 +10,6 @@ import {
 import { WebAPICallResult } from '@slack/web-api';
 import { SlackMessageRow } from 'knex/types/tables';
 
-import { getUser } from '@/api/getUser';
 import { bolt } from '@/api/slack';
 import * as slackblocks from '@/blocks/slackBlocks';
 import { GETSENTRY_REPO_SLUG, GH_ORGS, SENTRY_REPO_SLUG } from '@/config';
@@ -19,6 +18,7 @@ import { GoCDModification, GoCDPipeline, GoCDResponse } from '@/types';
 import { getLastGetSentryGoCDDeploy } from '@/utils/db/getLatestDeploy';
 import { getSlackMessage } from '@/utils/db/getSlackMessage';
 import { saveSlackMessage } from '@/utils/db/saveSlackMessage';
+import { getUser } from '@/utils/github/getUser';
 import {
   filterBuildCauses,
   firstGitMaterialSHA,
