+++ src/brain/gocdSlackFeeds/index.test.ts
@@ -16,7 +16,7 @@ import {
   GOCD_SENTRYIO_BE_PIPELINE_NAME,
 } from '@/config';
 import { Fastify, GoCDResponse } from '@/types';
-import { getUser as nonMockedGetUser } from '@api/getUser';
+import { getUser as nonMockedGetUser } from '@/utils/github/getUser';
 import { bolt as originalBolt } from '@api/slack';
 import { db } from '@utils/db';
 
@@ -28,7 +28,7 @@ import {
   ROLLBACK_PLAYBOOK_LINK,
 } from '.';
 
-jest.mock('@api/getUser');
+jest.mock('@/utils/github/getUser');
 const getUser = nonMockedGetUser as jest.Mock;
 
 describe('gocdSlackFeeds', function () {
