+++ src/brain/pleaseDeployNotifier/actionViewUndeployedCommits.ts
@@ -6,9 +6,9 @@ import {
   GOCD_SENTRYIO_BE_PIPELINE_GROUP,
   GOCD_SENTRYIO_BE_PIPELINE_NAME,
 } from '@/config';
+import { getRelevantCommit } from '@/utils/github/getRelevantCommit';
 import { firstGitMaterialSHA } from '@/utils/gocdHelpers';
-import { getBlocksForCommit } from '@api/getBlocksForCommit';
-import { getRelevantCommit } from '@api/github/getRelevantCommit';
+import { getBlocksForCommit } from '@/utils/slack/getBlocksForCommit';
 import { getLastGetSentryGoCDDeploy } from '@utils/db/getLatestDeploy';
 
 /**
