+++ src/brain/pleaseDeployNotifier/index.ts
@@ -5,7 +5,6 @@ import * as Sentry from '@sentry/node';
 import { KnownBlock } from '@slack/types';
 
 import { githubEvents } from '@/api/github';
-import { getChangedStack } from '@/api/github/getChangedStack';
 import { getUpdatedGoCDDeployMessage } from '@/blocks/getUpdatedDeployMessage';
 import { gocdDeploy } from '@/blocks/gocdDeploy';
 import { muteDeployNotificationsButton } from '@/blocks/muteDeployNotificationsButton';
@@ -20,13 +19,14 @@ import {
 } from '@/config';
 import { SlackMessage } from '@/config/slackMessage';
 import { getGoCDDeployForQueuedCommit } from '@/utils/db/getDeployForQueuedCommit';
+import { getChangedStack } from '@/utils/github/getChangedStack';
+import { getRelevantCommit } from '@/utils/github/getRelevantCommit';
+import { getUser } from '@/utils/github/getUser';
+import { isGetsentryRequiredCheck } from '@/utils/github/isGetsentryRequiredCheck';
 import { INPROGRESS_MSG, READY_TO_DEPLOY } from '@/utils/gocdHelpers';
-import { getBlocksForCommit } from '@api/getBlocksForCommit';
-import { getUser } from '@api/getUser';
-import { getRelevantCommit } from '@api/github/getRelevantCommit';
-import { isGetsentryRequiredCheck } from '@api/github/isGetsentryRequiredCheck';
+import { getBlocksForCommit } from '@/utils/slack/getBlocksForCommit';
+import { slackMessageUser } from '@/utils/slack/slackMessageUser';
 import { bolt } from '@api/slack';
-import { slackMessageUser } from '@api/slackMessageUser';
 import { saveSlackMessage } from '@utils/db/saveSlackMessage';
 import { wrapHandler } from '@utils/wrapHandler';
