+++ src/brain/requiredChecks/handleNewFailedBuild.ts
@@ -9,8 +9,8 @@ import { jobStatuses } from '@/blocks/jobStatuses';
 import { BuildStatus, Color, REQUIRED_CHECK_CHANNEL } from '@/config';
 import { SlackMessage } from '@/config/slackMessage';
 import { CHECK_RUN_PROPERTIES, CheckRun, CheckRunProperty } from '@/types';
-import { getBlocksForCommit } from '@api/getBlocksForCommit';
-import { getRelevantCommit } from '@api/github/getRelevantCommit';
+import { getRelevantCommit } from '@/utils/github/getRelevantCommit';
+import { getBlocksForCommit } from '@/utils/slack/getBlocksForCommit';
 import { bolt } from '@api/slack';
 import { getFailureMessages } from '@utils/db/getFailureMessages';
 import { saveSlackMessage } from '@utils/db/saveSlackMessage';
