+++ src/utils/github/getUser/index.test.ts
@@ -2,7 +2,7 @@ import { SLACK_PROFILE_ID_GITHUB } from '@/config';
 import { bolt } from '@api/slack';
 import { db } from '@utils/db';
 
-import { getUser } from './';
+import { getUser } from '.';
 
 describe('getUser', function () {
   beforeAll(async function () {
