+++ src/utils/slack/getBlocksForCommit/index.ts
@@ -2,7 +2,7 @@ import { KnownBlock } from '@slack/types';
 
 import { ReposGetCommit } from '@types';
 
-import { getUser } from '../getUser';
+import { getUser } from '../../github/getUser';
 
 type Options = {
   shouldSlackMention?: boolean;
